split handlers and stats to files
