+++ stats.go
@@ -1,0 +1,8 @@
+package statsviz
+
+import "runtime"
+
+type stats struct {
+	Mem          runtime.MemStats
+	NumGoroutine int
+}
